Refactoring to use shared JS for forms
